only run oxide tests instead of oxide and stable tests
We are currently working locally with 1 engine at a time. Either `oxide`
or `stable` not both. However, using `crosscheck` we currently run both
`oxide` AND `stable` tests in the `oxide` engine.

This worked up until now because the global `__OXIDE__` was modified at
runtime in the `crosscheck` function. However, the files on disk are
only pointing to a single engine implementation right now.

Note: CI will still `oxide` and `stable` engine tests, but it also
ensures that `node ./scripts/swap-engines.js` is executed. If you want
to run the tests of the _other_ engine locally then you also have to run
`node ./scripts/swap-engines.js`.